Small updates for clarity and Union size.
I've tried to clarify that the data should be entered in desired Hoist size in Feet. Which means that Metric users must decide the desired Hoist size in Feet prior to running the function. You can't enter the desired size in Metric and then have it give the overall metric dimensions. Perhaps this is something to work on later? In any case, I also added the dimensions of the Union (the star field). Things perhaps still to do (other than what was previously mentioned), is to show what the distance is between star centers in inches (or perhaps centimeters?) rather than (or in addition to) feet and meters.